Now have a working version of the macro machine generator that only generates the portions of the TM table actually used.  This is much faster and uses a lot less memory - yea!
In fact, it is so small that even with "unsigned long long" for the macro
symbols the computation runs out of bits for an otherwise feasible computation
- boo!

I'm looking into arbitrary precision integer libraries or storing the macro
symbols as, well, a sequence/array of symbols, ;-)...

+++ macro_machine_C.h
@@ -44,9 +44,9 @@ typedef struct macro_transition MACRO_TRANSITION;
 
 struct macro_transition
 {
-  int symbol;
+  unsigned long long symbol;
 
-  int w;
+  unsigned long long w;
   int d;
   int s;
 
@@ -61,14 +61,14 @@ typedef struct
 typedef struct
 {
   int num_states;
-  int num_symbols;
+  unsigned long long num_symbols;
 
   MACRO_STATE* machine;
 
   int* tape;
   int  tape_length;
 
-  int symbol;
+  unsigned long long symbol;
 
   int                total_symbols;
   unsigned long long total_steps;
@@ -79,7 +79,7 @@ typedef struct
 
   int state;
 
-  int new_symbol;
+  unsigned long long new_symbol;
   int new_delta;
   int new_state;
